No se permite la redimension de la ventana en modo windowed (pantalla NO completa)
Se cambia el HUD del juego para que se pueda ver bien en todas las resoluciones y las imagenes de él no pierdan el aspect ratio que deben tener.

+++ Src/Logic/Entity/Components/Shop.cpp
@@ -478,7 +478,7 @@ namespace Logic
 			basestring << "mapRed";
 
 			unsigned short int ring = _gameStatus->getPlayer(player)->getPlayer()->getLogicalPosition()->getRing();
-			unsigned short int degree =_gameStatus->getPlayer(player)->getPlayer()->getLogicalPosition()->getDegree();
+			unsigned short int degree = (unsigned short int) _gameStatus->getPlayer(player)->getPlayer()->getLogicalPosition()->getDegree();
 			unsigned short int sense = _gameStatus->getPlayer(player)->getPlayer()->getLogicalPosition()->getSense();
 
 			//correct sense if player is looking in another direction that LEFT or RIGHT
